Zip file I
Since the file was too heavy so I compressed it into a zip file and then uploaded it on GitHub. If anyone wants to have a look then download both the zip file and the train dataset and just update the directory address in the second cell of the notebook.